fixed type qualification to support nested dtos, virtual namespaces, namespaces on dtos, etc.

+++ apyxl/examples/csharp/User.cs
@@ -30,7 +30,7 @@ public class User
     public struct Id
     {
         // feature: property accessor
-        public UInt128 Value { get; }
+        public UInt64 Value { get; }
     }
     
     // feature: nested enum
